return as int

+++ server/src/controllers/plaid_controller.rs
@@ -31,7 +31,7 @@ pub struct AccountResponse {
 pub struct AccountSuccess {
   pub item_id: String,
   pub name: String,
-  pub balance: f64,
+  pub balance: i64,
 }
 
 #[derive(Serialize)]
